# docs/swagger.yaml
basePath: /
definitions:
  models.SecretRequest:
    properties:
      data:
        additionalProperties:
          type: string
        description: Arbitrary key-value pairs stored in the secret
        example:
          '"password"': '"s3cr3t"}'
          '{"username"': '"admin"'
        type: object
      secret-name:
        description: Name of the secret
        example: db-credentials
        type: string
    required:
    - data
    - secret-name
    type: object
  models.SecretResponse:
    properties:
      data:
        additionalProperties:
          type: string
        description: |-
          Key-value pairs stored in the secret
          example: {"username":"admin","password":"s3cr3t"}
        type: object
      secret-name:
        description: |-
          Name of the secret
          example: db-credentials
        type: string
    type: object
  models.UserRequest:
    properties:
      password:
        description: User password (plain text, will be hashed server-side)
        example: P@ssw0rd123
        type: string
      username:
        description: Username chosen by the user
        example: stefan
        type: string
    required:
    - password
    - username
    type: object
  models.UserResponse:
    properties:
      message:
        description: Human-readable message describing the result
        example: Login successful
        type: string
      token:
        description: JWT access token returned after successful login
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: stefan.todorov24@mail.bg
    name: Stefan Todorov
    url: https://github.com/Stefan956
  description: Secure API for managing secrets in Kubernetes
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Stefan956/secretsManagerAPI
  title: Secrets Manager API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns a JWT token
      parameters:
      - description: User credentials payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Invalid username or password
          schema:
            type: string
        "500":
          description: Failed to generate token
          schema:
            type: string
      summary: Authenticate user
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user namespace and stores credentials in a secret
      parameters:
      - description: New User credentials payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to create user
          schema:
            type: string
      summary: Register a new user
      tags:
      - User
  /secrets/create/:
    post:
      consumes:
      - application/json
      description: Creates a Kubernetes secret in the authenticated user's namespace
      parameters:
      - description: Secret payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SecretRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SecretResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new secret
      tags:
      - Secrets
  /secrets/delete/:
    delete:
      description: Deletes a secret by name
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Secret name missing
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a secret
      tags:
      - Secrets
  /secrets/get/:
    get:
      description: Retrieves a secret by name from the authenticated user's namespace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SecretResponse'
        "400":
          description: Secret name missing
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Secret not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a secret
      tags:
      - Secrets
  /secrets/update/:
    put:
      consumes:
      - application/json
      description: Updates an existing secret by name
      parameters:
      - description: Updated secret data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SecretResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a secret
      tags:
      - Secrets
  /user/change-password/:
    put:
      consumes:
      - application/json
      parameters:
      - description: New password payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to update password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - User
  /user/delete/:
    delete:
      description: Deletes the user namespace and all associated resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to delete user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Register, Login, Update and Delete User
  name: User
- description: CRUD REST APIs for Secrets
  name: Secrets